moved gifviewer to live within giphyblock
